requirements.txt
flask
flask_sqlalchemy
requests
flask_migrate
event_bus